removed zoom swinging, error prone and unpopular, other proposed fixes
