+++ src/gameData/GameConsts.h
@@ -24,13 +24,6 @@
 /// Geschwindigkeitsabstufungen - Längen der GFs in ms
 const boost::array<unsigned, 6> SUPPRESS_UNUSED SPEED_GF_LENGTHS = {{80, 60, 50, 40, 30, 1}};
 
-const boost::array<float, 5> ZOOM_FACTORS = { 0.5f, 0.75f, 1.f, 1.25f, 1.5f};
-const size_t ZOOM_DEFAULT_INDEX = 2;
-const float ZOOM_ACCELERATION = 0.001f;
-const float ZOOM_WHEEL_INCREMENT = 0.03f;
-const float ZOOM_MIN = 0.1f;
-const float ZOOM_MAX = 3.f;
-
 /// Konvertierungstabelle von Rohstoff-Indizes von den Bergwerken --> Map
 const boost::array<unsigned char, 5> SUPPRESS_UNUSED RESOURCES_MINE_TO_MAP = {{3, 0, 1, 2, 4}};
